Order will not be placed if stock has zero or lesser than that. Created Post endpoint to create inventory. Tested and working
Updated order-service and inventory-service with saga pattern
to rollback the order if the order has failed and display with message on
Postman and save the status in order-service DB. Tested and working

Added new columns brandName and modelNumber. Added Get endpoints with
Global Exceptions in both and order and inventory Services.
..Tested and working.

Updated JsonFiles for order and inventory services.

Signed-off-by: Chaitanya Yenamandra <chaitanya0368@gmail.com>