Commit #3 for customer-service
Added emailAddress, password columns.

Created unique customerId is response so that original Id is not exposed
in response.

encrypted password using BcryptPasswordEncoder.

Added DTO and Mapper classes so that data is not saved and retrieved directly
from the entity classed but it should be done using DTO and mapper classes.

Signed-off-by: Chaitanya Yenamandra <chaitanya0368@gmail.com>