Added AppConstants in order-service and Spring Gateway to centralize messages which are displayed in Controller class.
Implemented Authorization in Spring Gateway...so that
token is created on role basis. Classes have been added in Spring Gateway
and customer-service.

Updated OrderController to implement Hybrid approach
to get the live status of the inventory with RestTemplate(Synchronous check)
and rollback the status using kafka(Asynchronous Kafka-based compensation)
and show the response in postman and place the correct orderStatus in DB.

We are doing a Saga Pattern with enhancements so that we get user-friendly
REST+Kafka hybrid that gives real-time response.

Updated JsonEntries text document for all order, inventory and customer-service.

Signed-off-by: Chaitanya Yenamandra <chaitanya0368@gmail.com>